Refactor function names.
Remove "peak_finder_" prefix from function names as the context already
denotes this and the point in having more than one function for the same
objective is exactly to show differences in the approaches. Therefore
names following regular naming conventions (e.g., "get_peak()") are not
necessarily well-suited for this, while naming them according to the
strategy is.

Signed-off-by: Lucas Caparelli <lucas.caparelli112@gmail.com>